chore: Control flow wip
